+++ index.html
@@ -228,7 +228,7 @@
     </section>
   </template>
 
-  <script src="js/functions.js"></script>
-  <script src="main.js"></script>
+  <script src="js/functions.js" type="module"></script>
+  <script src="js/main.js"></script>
 </body>
 </html>
