+++ js/util.js
@@ -1,0 +1,18 @@
+const getRandomInt = (a,b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a,b));
+  return Math.floor(Math.random() * (upper - lower + 1)) + lower;
+};
+
+const getRandomArrayElement = (items) =>
+  items[getRandomInt(0, items.length - 1)];
+
+const createIdGen = () => {
+  let lastGeneratedId = 0;
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export { getRandomInt, getRandomArrayElement, createIdGen };
